Should fix #8

+++ Hormones/database-brief.sql
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS hormones_tissues (
 	maxSlots SMALLINT UNSIGNED,
 	ip VARCHAR(68),
 	port SMALLINT UNSIGNED,
-	hormonesVersion SMALLINT,
+	hormonesVersion MEDIUMINT,
 	displayName VARCHAR(100),
 	processId SMALLINT UNSIGNED,
 	FOREIGN KEY (organId) REFERENCES hormones_organs(organId) ON UPDATE CASCADE ON DELETE RESTRICT
